Move código comum para trait
